2 charge removal of stickies

+++ code/modules/xenomorph/xeno_structures.dm
@@ -600,9 +600,15 @@ TUNNEL
 
 	var/datum/effect_system/smoke_spread/xeno/acid/acid_smoke
 
-	for(var/obj/item/explosive/grenade/sticky/sticky_bomb in stepper.contents)
-		sticky_bomb.clean_refs()
-		qdel(sticky_bomb)
+	if(charges >= 2)
+		for(var/obj/item/explosive/grenade/sticky/sticky_bomb in stepper.contents)
+			sticky_bomb.clean_refs()
+			qdel(sticky_bomb)
+			if(isxeno(stepper))
+				charges--
+				charges--
+				update_icon()
+				return
 	if(isxeno(stepper))
 		if(!(stepper.on_fire))
 			return
